==> 02_Spatial_Temporal_Groups_Seasons.R <==
# Author: Tatiana Proboste & Abigail Turnlund
# date: 24/05/2023 / 01/11/2023

# we based this analysis on Robitaille et al 2019 study 
# using the R package spatsoc (https://doi.org/10.1111/2041-210X.13215)

#Libraries
library(tidyverse)
library(sf)
library(adehabitatLT)
library(spatsoc) # GPS - spatial temporal analysis 
library(asnipe)
library(tmap)

#load the data

df <- readRDS("data.RDS")
df <- as.data.frame(df)


df <- sf::st_as_sf(df,  coords=c('longitude','latitude'))
st_crs(df) <- 32617 # GDA2020

# No geometry in CTMM File
df <- df %>% 
  mutate(x = unlist(map(df$geometry,1)),
        y = unlist(map(df$geometry,2))) %>% 
  st_as_sf(df)

df <- df %>%
  dplyr::mutate(month = lubridate::month(datetime)) %>%
  dplyr::mutate(season = dplyr::case_when(
    month %in% c(12, 1, 2) ~ "Summer",
    month %in% c(3, 4, 5) ~ "Autumn",
    month %in% c(6, 7, 8) ~ "Winter",
    month %in% c(9, 10, 11) ~ "Spring"
  ))

# -------------------------------------------------------------------------

df <- data.table::setDT(df) %>% 
  dplyr::select(ID, population, yr, datetime, x, y, season)

# temporal group
df <- spatsoc::group_times(
  df, 
  datetime = 'datetime',
  threshold = '5 minutes') # '5 minutes' direct contact - '5 days' indirect contact

# Threshold ----------------------------------------------------
# Spatial group using 'timegroup generated by group_times 
### separate by 0.00315 degrees or ~350m 
### seperate by 0.000018 degrees or ~ 2m
### separate by 0.000045 degrees of ~ 5m

output <- group_pts(
  df,
  threshold = 0.000045,  # convert from meters to degrees 
  id = 'ID',
  coords =  c('x', 'y'),
  timegroup = 'timegroup',
  splitBy = c('population', 'yr', 'season'))


write_rds(output, "Output.rds")
